src/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
pyautogui>=0.9.54
